backend/requirements.txt
asgiref==3.5.2
channels==4.0.0
Django==4.1.3
djangorestframework==3.14.0
